migration

+++ crm-backend/accounts/models.py
@@ -17,7 +17,7 @@ class Product(models.Model):
     id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=200, null=True)
     price = models.FloatField(null=True)
-    quantity = models.FloatField(null=True)
+    quantity = models.IntegerField(null=True)
     description = models.CharField(max_length=200, null=True, blank=True)
     date_created = models.DateTimeField(auto_now_add=True, null=True)
